Consolidated agent files into one, utilize pythond standard libaries
