+++ agent/agent_config.json
@@ -1,7 +1,7 @@
 {
     "scan_agent": "agent1",
     "api_token": "",
-    "master_address": "https://127.0.0.1",
+    "master_address": "127.0.0.1",
     "master_port": 4430,
     "callback_interval_in_seconds": 60,
     "number_of_threads": 1,
