+++ agent/requirements.txt
@@ -1,5 +1,2 @@
 # ipython, not required, helpful when needing a Python shell.
 ipython
-
-# HTTP requests
-requests==2.22.0
